Robert/370 teachers booking full student event (#183)
* [370]: Add extra check to getPlacesAvailable, modify ensureCapacity to allow all non-students to ignore limits rather than just teachers

* [370]: Refactor getPlacesAvailable to reduce the cognitive complexity below linter limit

* [370]: New unit tests for getPlacesAvailable, cleaned up older tests for method

* [370]: Parameterize a couple of booking tests to check that other non-student roles are subject to the same rules as teachers on student events

* [370]: Change numberOfPlaces on event pages from Long to Integer

+++ src/main/java/uk/ac/cam/cl/dtg/isaac/dao/EventBookingPersistenceManager.java
@@ -151,7 +151,7 @@ public class EventBookingPersistenceManager {
    * @return Map of booking status, role to count
    * @throws SegueDatabaseException - if something is wrong with the database
    */
-  public Map<BookingStatus, Map<Role, Long>> getEventBookingStatusCounts(final String eventId,
+  public Map<BookingStatus, Map<Role, Integer>> getEventBookingStatusCounts(final String eventId,
                                                                          final boolean includeDeletedUsersInCounts)
       throws SegueDatabaseException {
     return dao.getEventBookingStatusCounts(eventId, includeDeletedUsersInCounts);
